pyproject.toml
[project]
name = "sqlagent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "langchain-deepseek>=0.1.3",
    "langchain-openai>=0.3.18",
    "python-dotenv>=1.1.0",
]
